pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GRiD"
version = "0.1.0"
description = "Genomic Repeat inference from Depth - LPA KIV-2 Copy Number Variant Estimation Pipeline"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["bioinformatics", "genomics", "copy-number-variation", "LPA", "KIV-2", "VNTR", "depth", "coverage"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pysam>=0.21.0",
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "jinja2>=3.0",
    "rich>=13.0",  # For beautiful CLI output
    "dataclasses-json>=0.5.7",
    "validators>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "isort>=5.12",
]

[project.scripts]
grid = "grid.cli:main"

[project.urls]
Homepage = "https://github.com/caterer-z-t/GRiD"
Documentation = "https://grid.readthedocs.io"
Repository = "https://github.com/caterer-z-t/GRiD"
Issues = "https://github.com/caterer-z-t/GRiD/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["grid*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true